make/tools: limit concurrent go build processes to 8
Reasoning: some of the projects are large enough that they cause the
system to run out of memory before they finish building. when limiting
concurrent build processes to 8, the project still builds in a
reasonably fast amount of time, but doesn't use up a significant amount
of memory, at least on my Ryzen 5900 and 16GB RAM system. i think given
the CPU speed is one of the fastest available right now for desktop
machines means that it's definitely going to be limited by CPU rather
than memory.

time to download more RAM